change int to long in tutorial 08

+++ 08_sum/sum.c
@@ -2,18 +2,18 @@
 
 int main() {
 
-    int n, sum = 0;
+    long n, sum = 0;
 
     printf("This program sum a series of integers.\n");
     printf("Enter integers (0 to terminates): ");
 
-    scanf("%d", &n);
+    scanf("%ld", &n);
     while (n != 0) {
         sum += n;
-        scanf("%d", &n);
+        scanf("%ld", &n);
     }
 
-    printf("The sum is: %d\n", sum);
+    printf("The sum is: %ld\n", sum);
 
     return 0;
 }
